created register email service for new users. Sends a welcome email to that user and will be used in the future to send survey links etc.
changed some template logic so that it doesn't crash the site if there 
are no meals to display -- however, still revamping the get_meals() 
function to be more dynamic e.g. what if there is only one meal? so, 
doesn't work yet.

in addition to that, linked the display of meals on the chef page with 
the user_id for that chef so that only the meals that chef created are 
displayed on the page.